Add files via upload
this is a discord game bot i was working on but there were so many bugs and i completely made a mess putting everything in one file and many times i was completely stcuk and if i wanted to repair it i had to completely rewrite the entire code. this was supposed to be the most fun discord bot to add in your server to play games it had so so many features and i planned more but it was too much for me alone, so i dropped it. if u can help me out or use this code as an sample u are free to use or help me out.